Refactoring the plugin. Moving the responsability for deploying away from ArtifactoryTask into a new DeployTask existing in the root project
