+++ build-info-extractor-gradle/src/main/groovy/org/jfrog/gradle/plugin/artifactory/ArtifactoryPlugin.groovy
@@ -4,8 +4,10 @@ import org.gradle.api.Project
 import org.jfrog.gradle.plugin.artifactory.dsl.ArtifactoryPluginConvention
 import org.jfrog.gradle.plugin.artifactory.task.ArtifactoryTask
 import org.jfrog.gradle.plugin.artifactory.task.BuildInfoBaseTask
+import org.jfrog.gradle.plugin.artifactory.task.DeployTask
 
 import static org.jfrog.gradle.plugin.artifactory.task.BuildInfoBaseTask.BUILD_INFO_TASK_NAME
+import static org.jfrog.gradle.plugin.artifactory.task.BuildInfoBaseTask.DEPLOY_TASK_NAME
 
 /**
  * @author Lior Hasson  
@@ -19,8 +21,15 @@ class ArtifactoryPlugin extends ArtifactoryPluginBase {
     @Override
     protected BuildInfoBaseTask createArtifactoryPublishTask(Project project) {
         def result = project.getTasks().create(BUILD_INFO_TASK_NAME, ArtifactoryTask.class)
-        result.setDescription('''Deploys artifacts + generated build-info metadata to Artifactory,
+        result.setDescription('''Add artifacts + generated build-info metadata to be deployed to Artifactory,
                                  using project configurations.''')
+        return result
+    }
+
+    @Override
+    protected DeployTask createArtifactoryDeployTask(Project project) {
+        def result = project.getTasks().create(DEPLOY_TASK_NAME, DeployTask.class)
+        result.setDescription('''Deploys all queued ArtifactoryTask''')
         return result
     }
 }
